refactor(frontend): Simplify model index
