+++ apps/frontend/src/app/model/index.ts
@@ -1,7 +1,19 @@
-export * from './participant';
-export * from './poll';
-export * from './poll-event';
-export * from './settings';
 export * from './token';
 export * from './statistics';
-export * from './read-poll';
+
+import * as types from '@apollusia/types';
+
+export type Poll = types.DTO<types.Poll>;
+export type CreatePollDto = Omit<Poll, '_id'>;
+
+export interface ReadPoll extends Poll {
+  isAdmin: boolean;
+  events: number;
+  participants: number;
+}
+
+export type PollEvent = types.DTO<types.PollEvent> & { participants?: Participant[]; };
+export type CreatePollEventDto = Omit<PollEvent, '_id' | 'poll'> & { _id?: string; };
+
+export type Participant = types.DTO<types.Participant>;
+export type CreateParticipantDto = Omit<Participant, '_id' | 'poll'>;
