--- apps/frontend/src/app/model/participant.ts
@@ -1,5 +1,0 @@
-import * as types from '@apollusia/types';
-
-export type Participant = types.DTO<types.Participant>;
-
-export type CreateParticipantDto = Omit<Participant, '_id' | 'poll'>;
