--- apps/frontend/src/app/model/poll.ts
@@ -1,5 +1,0 @@
-import * as types from '@apollusia/types';
-
-export type Poll = types.DTO<types.Poll>;
-
-export type CreatePollDto = Omit<Poll, '_id'>;
