--- apps/frontend/src/app/model/read-poll.ts
@@ -1,7 +1,0 @@
-import {Poll} from './poll';
-
-export interface ReadPoll extends Poll {
-  isAdmin: boolean;
-  events: number;
-  participants: number;
-}
