--- apps/frontend/src/app/model/settings.ts
@@ -1,4 +1,0 @@
-import * as types from '@apollusia/types';
-
-
-export type Settings = types.DTO<types.Settings>;
